Service : Security check on CRUD routes in front to check if user has ROLE_ADMIN

+++ src/Controller/ServiceController.php
@@ -27,6 +27,16 @@ class ServiceController extends AbstractController
      */
     public function create(Request $request): Response
     {
+        // MODIFS SEB - START
+        // Permet de vérifier si utilisateur connecté et si c'est un admin.
+        // Si ce n'est pas un admin alors on le redirige vers l'index des services car il n'a rien à faire ici.
+        // => Code copié collé sur les 3 routes du crud : create/edit/delete
+        $user = $this->getUser();
+        if(!$user OR !in_array('ROLE_ADMIN', $user->getRoles())) {
+            return $this->redirectToRoute('services_index');
+        } 
+        // MODIFS SEB - END
+        
         $service = new Service();
 
         $form = $this->createForm(ServiceType::class, $service);
@@ -50,6 +60,16 @@ class ServiceController extends AbstractController
      */
     public function edit(Request $request, Service $service)
     {
+        // MODIFS SEB - START
+        // Permet de vérifier si utilisateur connecté et si c'est un admin.
+        // Si ce n'est pas un admin alors on le redirige vers l'index des services car il n'a rien à faire ici.
+        // => Code copié collé sur les 3 routes du crud : create/edit/delete
+        $user = $this->getUser();
+        if(!$user OR !in_array('ROLE_ADMIN', $user->getRoles())) {
+            return $this->redirectToRoute('services_index');
+        } 
+        // MODIFS SEB - END
+        
         $form = $this->createForm(ServiceType::class, $service);
         $form->handleRequest($request);
 
@@ -68,6 +88,16 @@ class ServiceController extends AbstractController
      */
     public function delete($id, ServiceRepository $serviceRepository)
     {
+        // MODIFS SEB - START
+        // Permet de vérifier si utilisateur connecté et si c'est un admin.
+        // Si ce n'est pas un admin alors on le redirige vers l'index des services car il n'a rien à faire ici.
+        // => Code copié collé sur les 3 routes du crud : create/edit/delete
+        $user = $this->getUser();
+        if(!$user OR !in_array('ROLE_ADMIN', $user->getRoles())) {
+            return $this->redirectToRoute('services_index');
+        } 
+        // MODIFS SEB - END
+        
         $em = $this->getDoctrine()->getManager();
         $service = $serviceRepository->find($id);
         $em->remove($service);
